tried eitans suggest. still weird results.
The strange thing that I would focus on now is that the timings for 2^21
through 2^24 have very little variance, relative to the other array
sizes.

I would expect them to have more variance, since the data is being
stored in different caches, so the times should vary. Or at the very
least be higher if the times dont' average out.